+++ trunk/cpsycle/audio/src/sequencer.h
@@ -1,5 +1,5 @@
 // This source is free software ; you can redistribute it and/or modify it under the terms of the GNU General Public License as published by the Free Software Foundation ; either version 2, or (at your option) any later version.
-// copyright 2000-2019 members of the psycle project http://psycle.sourceforge.net
+// copyright 2000-2020 members of the psycle project http://psycle.sourceforge.net
 
 #if !defined(SEQUENCER_H)
 #define SEQUENCER_H
@@ -66,7 +66,7 @@ typedef struct {
 void sequencer_init(psy_audio_Sequencer*, psy_audio_Sequence*, psy_audio_Machines*);
 void sequencer_dispose(psy_audio_Sequencer*);
 void sequencer_reset(psy_audio_Sequencer*, psy_audio_Sequence*, psy_audio_Machines*);
-void sequencer_frametick(psy_audio_Sequencer*, unsigned int numsamples);
+void sequencer_frametick(psy_audio_Sequencer*, uintptr_t numsamples);
 void sequencer_tick(psy_audio_Sequencer*, psy_dsp_beat_t offset);
 void sequencer_linetick(psy_audio_Sequencer*);
 void sequencer_setposition(psy_audio_Sequencer*, psy_dsp_beat_t position);
@@ -74,8 +74,9 @@ psy_dsp_beat_t sequencer_position(psy_audio_Sequencer*);
 void sequencer_start(psy_audio_Sequencer*);
 void sequencer_stop(psy_audio_Sequencer*);
 psy_List* sequencer_tickevents(psy_audio_Sequencer*);
-psy_List* sequencer_machinetickevents(psy_audio_Sequencer*, size_t slot);
-psy_List* sequencer_timedevents(psy_audio_Sequencer*, size_t slot, unsigned int amount);
+psy_List* sequencer_machinetickevents(psy_audio_Sequencer*, uintptr_t slot);
+psy_List* sequencer_timedevents(psy_audio_Sequencer*, uintptr_t slot,
+	uintptr_t amount);
 void sequencer_append(psy_audio_Sequencer*, psy_List* events);
 void sequencer_setsamplerate(psy_audio_Sequencer*, unsigned int samplerate);
 unsigned int sequencer_samplerate(psy_audio_Sequencer*);
@@ -86,9 +87,11 @@ uintptr_t sequencer_lpb(psy_audio_Sequencer*);
 unsigned int sequencer_frames(psy_audio_Sequencer*, psy_dsp_beat_t offset);
 psy_dsp_beat_t sequencer_frametooffset(psy_audio_Sequencer*, int numsamples);
 int sequencer_playing(psy_audio_Sequencer*);
-void sequencer_addinputevent(psy_audio_Sequencer*, const psy_audio_PatternEvent*, uintptr_t track);
-void sequencer_recordinputevent(psy_audio_Sequencer*, const psy_audio_PatternEvent*,
-	unsigned int track, psy_dsp_beat_t playposition);
+void sequencer_addinputevent(psy_audio_Sequencer*,
+	const psy_audio_PatternEvent*, uintptr_t track);
+void sequencer_recordinputevent(psy_audio_Sequencer*,
+	const psy_audio_PatternEvent*, uintptr_t track,
+	psy_dsp_beat_t playposition);
 void sequencer_setplaymode(psy_audio_Sequencer*, SequencerPlayMode);
 void sequencer_loop(psy_audio_Sequencer*);
 void sequencer_stoploop(psy_audio_Sequencer*);
